according to: http://www.gnu.org/software/autogen/manual/html_node/autogen_191.html#SEC191
The proper way to librarize a libopts library is to use a stub and seperate
the opts into a seperate .def which you can then include.  So I renamed
src/tcpedit_opts.[ch] to src/tcpedit/parse_args.[ch]

Things are kinda broken right now (again) but hopefully it won't be too much
work to fix.