Reverted some changes in main netmap loop to increase performance. - #106 Added 'netmap:foo!' for all queues By default only write to first queue - other queues are slow General fixes to interface names.
Note: Should review with Luigi. Under Linux, slots are not seen as available
unless NIOCTXSYNC is run after poll. Also, running NIOCTXSYNC seems to make
things faster. We can just TX as fast as possible, but a rediculously high
number of packets are reported sent (e.g. reports of 500Gbps). To get
accuracy, need to use NIOCTXSYNC.

Also, only queue 0 is fast on Linux. This is the default when using interface
'netmap:foo'. If using all queuse with 'netmap:foo!' we slow down. Sending to
'netmap:foo-0' is faster but not as fast as default. 'netmap:foo-7' is very
slow.