Add data fuzzing option to tcpreplay-edit
This fuzzing was designed as to test high-level protocols. It modifies
randomly 1 out of 8 packets in order for stateful protocols to cover
more of their code. The random fuzzing actions focus on data start and
end because it often is the part of the data high-level protocols base
their decisions on.

0x00 and 0xff are considered bytes of a higher significance than the
other values, and given more appearance probability.

When a packet is fuzzed, we randomly apply one action (see fuzzing.h)
If no action corresponds, FUZZING_CHANGE_MID_RANDOM is default fuzzing
case.

Option is named --fuzz-seed (without short option), not to be mixed with
--seed (randomize src/dst IPv4/v6 addresses).