Bug #474 enable pcap_findalldevs when --with-testnic is specified
This setting was being skipped when --with-testnic=xxx was specified to
configure.

Also, enables osx_frameworks var when the option is
specified when building for OS X.  This should make the option have only
the intended effect on the build.

There may be some other issues with the current structure.  E.g. lines
such as AC_DEFINE([HAVE_LINUX], [1], [Building Linux]) are never reached
with testnic defined.