Simplify plugin Makefiles
Add new TCPEDIT_PLUGINS_DEPS variable to the plugin's folder Makefile
which lists the files required for all plugins. Use that new variable in
all the plugins.
Since the plugin Makefiles are included from and after that variable
declaration, all the plugins hae access to it.

It also removes a double-reference to dlt_utils.h in many plugins