Bug #557 Attempt to correct corrupt packets
If packet caplen > packet len, update caplen to equal len.

Although this appears to be impossible if trace was generated from
a reliable source (i.e. libpcap), there appears to be enough pressure
from the community to allow these invalid packets to be sent.

IMHO, I think that if a pcap is invalid, tcpreplay should fail. But it
appears that this is the behavior desired by users.