send_packet: Avoid clock drift by using time since first packet
Prior to this change, time between packets was calculated as deltas
and over time errors would quickly add up. The work here switches
to use the wall time since the first packet was sent as the reference
which is compared to the PCAP packet timestamp relative to the first
packet. In this way there are no errors and clock drift is avoided.
As a result tcpreplay can now be used to play back very time sensitive
captures such as video streams with PCR.